waveshift-tts-engine/requirements.txt
aiofiles
Cython
einops
huggingface_hub==0.27.1
HyperPyYAML
joblib
json_repair
m3u8
numpy
omegaconf
openai
packaging
pandas
pesq==0.0.4
phonemizer
piper_phonemize
protobuf==4.25
pyarrow
pysubs2
python-dotenv
pyworld
PyYAML
Requests
scipy
setuptools
tiktoken
torchinfo
tqdm
transformers
Unidecode
wget
yamlargparse
google.generativeai
fastapi
uvicorn
rotary_embedding_torch
pydub
boto3
httpx
audio-separator[gpu]>=0.16.0